requirements.txt
alembic==1.9.4
anyio==3.6.2
anytree==2.8.0
apparmor==3.1.3
apsw==3.40.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autocommand==2.2.2
Babel==2.11.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.2
bleach==6.0.0
blinker==1.5
Brotli==1.0.9
brotlicffi==1.0.9.2
btrfsutil==6.2.2
cchardet==2.1.7
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
cryptography==39.0.2
css-parser==1.0.8
cssselect==1.2.0
Cython==0.29.33
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
distro==1.8.0
dnspython==2.3.0
docopt==0.6.2
email-validator==1.3.1
executing==1.2.0
extras==1.0.0
fastjsonschema==2.16.2
feedparser==6.0.10
filelock==3.8.2
fixtures==4.0.1
Flask==2.2.2
Flask-BabelEx==0.9.4
Flask-Compress==1.13
Flask-Gravatar==0.5.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==4.1.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
fqdn==1.5.1
gssapi==1.8.2
html2text==2020.1.16
html5-parser==0.4.10
html5lib==1.1
icu==0.0.1
idna==3.4
ifaddr==0.2.0
importlib-metadata==5.0.0
inflate64==0.3.1
inflect==6.0.2
ipykernel==6.20.2
ipython==8.9.0
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.context==4.3.0
jaraco.functools==3.6.0
jaraco.text==3.11.1
jarowinkler==1.2.3
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.0.2
jupyter_core==5.2.0
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.19.0
keyutils==0.6
ldap3==2.9
lensfun==0.3.3
Levenshtein==0.20.9
LibAppArmor==3.1.3
libtorrent==2.0.8
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mechanize==0.4.8
meson==1.0.1
mistune==2.0.4
more-itertools==9.1.0
mpmath==1.2.1
msgpack==1.0.4
multivolumefile==0.2.3
nbclassic==0.5.1
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
netsnmp-python==1.0a1
networkx==3.0
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.2
ordered-set==4.1.0
packaging==23.0
pandocfilters==1.5.0
paramiko==2.11.1
parso==0.8.3
passlib==1.7.4
pbr==5.11.1
pdftotext==2.2.2
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.1
Pillow==9.4.0
platformdirs==2.6.0
ply==3.11
polyglot==16.7.4
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
psycopg2==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pvporcupine==2.2.0
py7zr==0.20.2
pyasn1==0.4.8
PyAudio==0.2.13
pybcj==1.0.1
pycairo==1.23.0
pychm==0.8.6
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.12.0
pydantic==1.10.7
Pygments==2.14.0
PyGObject==3.44.1
pyinotify==0.9.6
PyNaCl==1.4.0
pyppmd==1.0.0
PyQt5==5.15.9
PyQt5-sip==12.11.1
PyQt6==6.4.2
PyQt6-sip==13.4.1
PyQt6-WebEngine==6.4.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-editor==1.0.4
python-json-logger==2.0.4
pyttsx3==2.90
pytz==2022.7.1
pyxdg==0.28
PyYAML==6.0
pyzmq==25.0.0
pyzstd==0.15.3
rapidfuzz==2.13.7
regex==2023.3.23
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.0
sgmllib3k==1.0.0
simplejson==3.18.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
speaklater==1.3
SQLAlchemy==1.3.24
sqlparse==0.4.3
srt==3.5.3
sshtunnel==0.4.0
stack-data==0.6.2
stevedore==4.1.1
sympy==1.11.1
team==1.0
terminado==0.17.1
testtools==2.6.0
texttable==1.6.7
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.0+cpu
torchaudio==2.0.1+cpu
torchvision==0.15.1+cpu
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
trove-classifiers==2023.3.10
typing_extensions==4.5.0
udiskie==2.4.2
ufw==0.36.1
ulauncher==5.15.0
unrardll==0.1.5
uri-template==1.2.0
urllib3==1.26.14
validate-pyproject==0.12.2.post1.dev0+g2940279.d20230328
virtualenv==20.17.1
virtualenv-clone==0.5.7
vosk==0.3.45
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.2
Werkzeug==2.2.3
WTForms==3.0.1
yt-dlp==2023.3.4
zeroconf==0.39.4
zipp==3.15.0
